fix(example-app): Remove dependencies to deprecated @angular/flex-layout
Usage of @angular/flex-layout was creating Conflicting peer dependency.
Removing the dependacy allowed to execute *npm install*
Removing usage from example-app allowed it to build.
No other change have been made to the html, so may not represent the intended output, but the librairy generaly work.
Since it's an app that showcase the librairie, it's acceptable.